Add files via upload
The application should contain 2 types of users: Hospitals and Receivers.
this a real-life system, that the real users will use. Hospital ‘Add blood info’ page - Once logged in, a hospital should be able to add details of available
blood samples (along with type) to their bank. Access to this page should be restricted only to the respective hospitals only.

Available blood samples page - A page should display all the available blood samples,
which hospital has them, and a ‘Request Sample’ button. This page should be accessible to
everyone, irrespective of whether the user is logged in or not.


 i. Only receivers should be able to request blood samples by clicking the ‘Request Sample’
button. Make sure that only those receivers who are eligible for the blood sample are allowed to
click the button.
ii. If the user is not logged in, then he/she should be redirected to the login page.
iii. If a user is logged in as a hospital, then the user should not be allowed to request a blood
sample.

Hospital ‘View requests’ page - Hospitals should be able to see the list of all the receivers who have
requested for a particular blood group from its blood bank.